backend/requirements.txt
django
psycopg2-binary
requests
python-dotenv
gunicorn
boto3
djangorestframework-simplejwt
pytz
channels
channels-redis 
daphne
djangorestframework
django-redis
openpyxl
django-cors-headers
deep-translator
cryptography
